Fix bug when parsing SpeakerNotes
