+++ PowerPointParser/PowerPointParserTests/Html/HtmlBuilderTests.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
-using System.IO;
 using Aaks.PowerPointParser.Dto;
 using FluentAssertions;
 using PowerPointParserTests.Html;
@@ -27,7 +26,7 @@ namespace Aaks.PowerPointParser.Html.Tests
 
             var actual = _htmlConverter.ConvertOpenXmlParagraphWrapperToHtml(wrapperList);
 
-            Assert.IsNull(actual);
+            actual.Should().BeEmpty(); 
         }
 
         [TestMethod]
@@ -39,7 +38,7 @@ namespace Aaks.PowerPointParser.Html.Tests
 
             var actual = _htmlConverter.ConvertOpenXmlParagraphWrapperToHtml(queue);
 
-            Assert.IsNull(actual);
+            actual.Should().BeEmpty();
         }
 
         [TestMethod]
@@ -51,7 +50,7 @@ namespace Aaks.PowerPointParser.Html.Tests
 
             var actual = _htmlConverter.ConvertOpenXmlParagraphWrapperToHtml(queue);
 
-            Assert.IsNull(actual);
+            actual.Should().BeEmpty();
         }
 
         [TestMethod]
@@ -67,7 +66,7 @@ namespace Aaks.PowerPointParser.Html.Tests
 
             var actual = _htmlConverter.ConvertOpenXmlParagraphWrapperToHtml(queue);
 
-            Assert.IsNull(actual);
+            actual.Should().BeEmpty(); 
         }
 
         [TestMethod]
@@ -449,29 +448,6 @@ namespace Aaks.PowerPointParser.Html.Tests
 
             Assert.AreEqual("<ol><li>one</li><li>two</li><ul><li>hello world</li><li>goodbye world</li><li>test world</li></ul><li>three</li></ol>", actual);
         }
-        [TestMethod]
-        public void Test_ParserWithWrapper()
-        {
-
-            var directory = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location) ?? string.Empty;
-            var filePath = Path.Combine(directory, "TestDeckParagraph.pptx");
-            File.Exists(filePath).Should().Be(true);
-            
-            var parser = new Parser();
-
-            var items = parser.ParseSpeakerNotes(filePath);
-
-            var innerBuilder = new InnerHtmlBuilder();
-
-            var htmlBuilder = new HtmlBuilder(new HtmlListBuilder(innerBuilder), innerBuilder);
-            var openXmlParagraphWrappers = items!.ToQueue();
-
-
-            var htmlStringActual = htmlBuilder.ConvertOpenXmlParagraphWrapperToHtml(openXmlParagraphWrappers!);
-            const string expected = @"<p>This note is just a paragraph</p><p>This note is just a paragraph</p><p>And this is a second paragraph</p><p><strong>This is a bold paragraph</strong></p><li>Unordered item 1</li><li>Unordered item 2</li><li>Unordered item 3</li><li>Indent Unordered item 1</li><ul><li>Indent Unordered item 2</li><ul><li>Indent Unordered item 3</li></ul></ul></ul><ol><li>Ordered one</li><li>Ordered two</li><li>Ordered three</li><li>Indent Ordered One</li><ol><li>Indent Ordered One One</li><ol><li>Indent Order One OneOne</li></ol></ol><li>Indent Ordered Three</li><p>Here a link: https://www.google.com/</p><li>Un</li><li>Order</li><li>List</li></ul><ol><li>Followed </li><li>by </li><li>Ordered</li></ol>";
-
-            htmlStringActual.Should().NotBeEmpty();
-            htmlStringActual.Should().Be(expected);
-        }
+       
     }
 }
