+++ PowerPointParser/PowerPointParserTests/PowerPointParserTests.csproj
@@ -8,6 +8,25 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Mocks\**" />
+    <Compile Remove="SampleOutput\**" />
+    <EmbeddedResource Remove="Mocks\**" />
+    <EmbeddedResource Remove="SampleOutput\**" />
+    <None Remove="Mocks\**" />
+    <None Remove="SampleOutput\**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Remove="TestData\TestThree.pptx" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="TestData\TestThree.pptx">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="6.5.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
     <PackageReference Include="Moq" Version="4.17.1" />
@@ -18,11 +37,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PowerPointParser\PowerPointParser.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Mocks\" />
-    <Folder Include="SampleOutput\" />
   </ItemGroup>
 
   <ItemGroup>
